feat: 添加用户认证与凭证管理功能
feat(api): 实现用户注册、登录和认证功能
feat(credential): 添加模型凭证的CRUD和验证功能
feat(audit): 引入审计日志记录关键操作
feat(project): 增强项目服务支持权限检查
feat(workflow): 完善工作流状态管理和时间戳跟踪

fix: 修复配置验证和模板渲染问题
fix(token): 修正令牌计数器的精度处理
fix(model): 更新审计日志和令牌使用模型

refactor: 重构服务工厂和依赖注入
refactor(context): 优化上下文构建器初始化
refactor(error): 统一错误处理中间件

test: 添加用户认证和凭证管理的单元测试
test: 补充工作流和服务层的测试用例

+++ apps/api/app/modules/billing/models/token_usage.py
@@ -1,6 +1,7 @@
+from decimal import Decimal
 import uuid
 
-from sqlalchemy import Float, ForeignKey, Integer, String
+from sqlalchemy import ForeignKey, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.shared.db.base import Base, TimestampMixin, UUIDMixin
@@ -20,4 +21,4 @@ class TokenUsage(Base, TimestampMixin, UUIDMixin):
     model_name: Mapped[str] = mapped_column(String(100))
     input_tokens: Mapped[int] = mapped_column(Integer)
     output_tokens: Mapped[int] = mapped_column(Integer)
-    estimated_cost: Mapped[float] = mapped_column(Float)
+    estimated_cost: Mapped[Decimal] = mapped_column(Numeric(12, 6))
